fix: hub push integration test
