+++ tests/integration/hub_usage/test_hub_usage.py
@@ -3,16 +3,26 @@ import subprocess
 
 import pytest
 
+from pathlib import Path
+
 from jina import __version__ as jina_version
 from jina.docker.hubio import HubIO
-from jina.excepts import PeaFailToStart, HubBuilderError
+from jina.docker import hubapi
+from jina.excepts import PeaFailToStart, HubBuilderError, ImageAlreadyExists
 from jina.executors import BaseExecutor
 from jina.flow import Flow
-from jina.parser import set_pod_parser, set_hub_build_parser
+from jina.jaml import JAML
+from jina.helper import expand_dict
+from jina.parser import set_pod_parser, set_hub_build_parser, set_hub_list_parser
 from jina.peapods import Pod
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 
+
+@pytest.fixture
+def access_token_github():
+    token = os.environ.get('GITHUB_TOKEN', None)
+    return token
 
 def test_simple_use_abs_import_shall_fail():
     with pytest.raises(ModuleNotFoundError):
@@ -76,3 +86,71 @@ def test_build_timeout_ready():
     HubIO(args).build()
     with Flow().add(uses=f'jinahub/pod.crafter.dummyhubexecutorslow:0.0.0-{jina_version}', timeout_ready = 20000):
         pass
+
+
+@pytest.mark.skipif(condition='os.environ.get(\'GITHUB_TOKEN\')==None', reason='Token not found')
+def test_hub_build_push(monkeypatch, access_token_github):
+    monkeypatch.setattr(Path, 'is_file', True)
+    monkeypatch.setattr(hubapi, '_fetch_access_token', access_token_github)
+    args = set_hub_build_parser().parse_args([str(cur_dir + 'dummyhub'), '--push', '--host-info'])
+    summary = HubIO(args).build()
+
+    with open(cur_dir + 'dummyhub' + 'manifest.yml') as fp:
+        manifest_jaml = JAML.load(fp, substitute=True)
+        manifest = expand_dict(manifest_jaml)
+
+    assert summary['is_build_success']
+    assert manifest['version'] == summary['version']
+    assert manifest['description'] == summary['manifest_info']['description']
+    assert manifest['author'] == summary['manifest_info']['author']
+    assert manifest['kind'] == summary['manifest_info']['kind']
+    assert manifest['type'] == summary['manifest_info']['type']
+    assert manifest['vendor'] == summary['manifest_info']['vendor']
+    assert manifest['keywords'] == summary['manifest_info']['keywords']
+
+    args = set_hub_list_parser().parse_args([
+        '--name', summary['manifest_info']['name'],
+        '--keywords', summary['manifest_info']['keywords'][0],
+        '--type', summary['manifest_info']['type']
+    ])
+    response = HubIO(args).list()
+    manifests = response
+
+    assert len(manifests) >= 1
+    assert manifests[0]['name'] == summary['manifest_info']['name']
+
+
+def test_hub_build_push_push_again(monkeypatch, access_token_github):
+    monkeypatch.setattr(Path, 'is_file', True)
+    monkeypatch.setattr(hubapi, '_fetch_access_token', access_token_github)
+    args = set_hub_build_parser().parse_args([str(cur_dir) + '/dummyhub', '--push', '--host-info'])
+    summary = HubIO(args).build()
+
+    with open(str(cur_dir) + '/dummyhub' + '/manifest.yml') as fp:
+        manifest_jaml = JAML.load(fp, substitute=True)
+        manifest = expand_dict(manifest_jaml)
+
+    assert summary['is_build_success']
+    assert manifest['version'] == summary['version']
+    assert manifest['description'] == summary['manifest_info']['description']
+    assert manifest['author'] == summary['manifest_info']['author']
+    assert manifest['kind'] == summary['manifest_info']['kind']
+    assert manifest['type'] == summary['manifest_info']['type']
+    assert manifest['vendor'] == summary['manifest_info']['vendor']
+    assert manifest['keywords'] == summary['manifest_info']['keywords']
+
+    args = set_hub_list_parser().parse_args([
+        '--name', summary['manifest_info']['name'],
+        '--keywords', summary['manifest_info']['keywords'][0],
+        '--type', summary['manifest_info']['type']
+    ])
+    response = HubIO(args).list()
+    manifests = response
+
+    assert len(manifests) >= 1
+    assert manifests[0]['name'] == summary['manifest_info']['name']
+
+    with pytest.raises(ImageAlreadyExists):
+        # try and push same version again should fail with `--no-overwrite`
+        args = set_hub_build_parser().parse_args([str(cur_dir) + '/dummyhub', '--push', '--host-info', '--no-overwrite'])
+        HubIO(args).build()
