+++ tests/unit/docker/test_hub_build.py
@@ -45,73 +45,6 @@ def test_hub_build_uses():
     # build again it shall not fail
     HubIO(args).build()
 
-@pytest.mark.skipif(condition='os.environ.get(\'GITHUB_TOKEN\')==None', reason='Token not found')
-def test_hub_build_push(monkeypatch, access_token_github):
-    monkeypatch.setattr(Path, 'is_file', True)
-    monkeypatch.setattr(hubapi, '_fetch_access_token', access_token_github)
-    args = set_hub_build_parser().parse_args([str(cur_dir / 'hub-mwu'), '--push', '--host-info'])
-    summary = HubIO(args).build()
-
-    with open(cur_dir / 'hub-mwu' / 'manifest.yml') as fp:
-        manifest_jaml = JAML.load(fp, substitute=True)
-        manifest = expand_dict(manifest_jaml)
-
-    assert summary['is_build_success']
-    assert manifest['version'] == summary['version']
-    assert manifest['description'] == summary['manifest_info']['description']
-    assert manifest['author'] == summary['manifest_info']['author']
-    assert manifest['kind'] == summary['manifest_info']['kind']
-    assert manifest['type'] == summary['manifest_info']['type']
-    assert manifest['vendor'] == summary['manifest_info']['vendor']
-    assert manifest['keywords'] == summary['manifest_info']['keywords']
-
-    args = set_hub_list_parser().parse_args([
-        '--name', summary['manifest_info']['name'],
-        '--keywords', summary['manifest_info']['keywords'][0],
-        '--type', summary['manifest_info']['type']
-    ])
-    response = HubIO(args).list()
-    manifests = response
-
-    assert len(manifests) >= 1
-    assert manifests[0]['name'] == summary['manifest_info']['name']
-
-
-def test_hub_build_push_push_again(monkeypatch, access_token_github):
-    monkeypatch.setattr(Path, 'is_file', True)
-    monkeypatch.setattr(hubapi, '_fetch_access_token', access_token_github)
-    args = set_hub_build_parser().parse_args([str(cur_dir / 'hub-mwu'), '--push', '--host-info'])
-    summary = HubIO(args).build()
-
-    with open(cur_dir / 'hub-mwu' / 'manifest.yml') as fp:
-        manifest_jaml = JAML.load(fp, substitute=True)
-        manifest = expand_dict(manifest_jaml)
-
-    assert summary['is_build_success']
-    assert manifest['version'] == summary['version']
-    assert manifest['description'] == summary['manifest_info']['description']
-    assert manifest['author'] == summary['manifest_info']['author']
-    assert manifest['kind'] == summary['manifest_info']['kind']
-    assert manifest['type'] == summary['manifest_info']['type']
-    assert manifest['vendor'] == summary['manifest_info']['vendor']
-    assert manifest['keywords'] == summary['manifest_info']['keywords']
-
-    args = set_hub_list_parser().parse_args([
-        '--name', summary['manifest_info']['name'],
-        '--keywords', summary['manifest_info']['keywords'][0],
-        '--type', summary['manifest_info']['type']
-    ])
-    response = HubIO(args).list()
-    manifests = response
-
-    assert len(manifests) >= 1
-    assert manifests[0]['name'] == summary['manifest_info']['name']
-
-    with pytest.raises(ImageAlreadyExists):
-        # try and push same version again should fail with `--no-overwrite`
-        args = set_hub_build_parser().parse_args([str(cur_dir / 'hub-mwu'), '--push', '--host-info', '--no-overwrite'])
-        HubIO(args).build()
-
 
 def test_hub_build_failures():
     for j in ['bad-dockerfile', 'bad-pythonfile', 'missing-dockerfile', 'missing-manifest']:
